Code/docqa/requirements.txt
# Web app
fastapi==0.135.1
gradio==5.49.1
python-multipart==0.0.22
uvicorn==0.41.0

# Configuration and data models
pydantic==2.11.5
python-dotenv==1.0.1

# LLM providers
google-generativeai==0.7.2
openai==2.26.0

# Document processing
numpy==1.26.4
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
pypdf==6.8.0
requests==2.32.5

# Retrieval and embeddings
llama-index-core==0.14.17
llama-index-embeddings-huggingface==0.7.0
llama-index-llms-gemini==0.6.2
llama-index-llms-openai==0.6.26
llama-index-workflows==2.16.1
nltk==3.9.3
sentence-transformers==3.0.1
tiktoken==0.12.0
